feat: add stream engine
